# backend/src/openapi/openapi.yaml
openapi: "3.1.0"
info:
  title: Coxa Fan OS API
  version: "0.2.0"
  description: |
    REST API for the Coxa Fan Operating System.
    Covers authentication, retail POS, ticketing, fan CDP, loyalty, and personalization.
    ## Authentication
    Obtain a JWT from POST /auth/login then pass Authorization: Bearer token.
    ## Tenant Context
    Every request needs x-tenant-id header (club identifier, e.g. coxa-club-001).
    ## Rate Limits
    Auth: 30/15min | Ticket ops: 60/min | Default: 500/min.
servers:
  - url: /api/v1
    description: Current API version
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    tenantHeader:
      type: apiKey
      in: header
      name: x-tenant-id
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
    User:
      type: object
      properties:
        _id:
          type: string
        fullName:
          type: string
        email:
          type: string
          format: email
        status:
          type: string
          enum: [active, inactive, suspended]
        createdAt:
          type: string
          format: date-time
    Club:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        slug:
          type: string
        country:
          type: string
        city:
          type: string
        tenantId:
          type: string
...
security:
  - bearerAuth: []
  - tenantHeader: []
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: Staff login
      description: Authenticates a staff user and returns a Bearer JWT plus club/membership info.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: admin@clubfc.com
                password:
                  type: string
                  format: password
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      user:
                        ref: "#/components/schemas/User"
                      club:
                        ref: "#/components/schemas/Club"
                      membership:
                        type: object
                        properties:
                          role:
                            type: string
                      token:
                        type: string
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                ref: "#/components/schemas/Error"
  /auth/signup:
    post:
      tags: [Authentication]
      summary: Create club account
      description: Creates a new staff user account plus club organization in one step.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fullName, email, password, clubName, country, city]
              properties:
                fullName:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                clubName:
                  type: string
                country:
                  type: string
                city:
                  type: string
      responses:
        "201":
          description: Account created
        "409":
          description: Email already in use
  /auth/fan/login:
    post:
      tags: [Authentication]
      summary: Fan login
      description: Authenticates a fan user. Returns JWT and fan profile.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Fan login successful
        "403":
          description: No fan profile linked to this account
  /auth/fan/signup:
    post:
      tags: [Authentication]
      summary: Fan registration
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fullName, email, password]
              properties:
                fullName:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                phone:
                  type: string
                memberId:
                  type: string
      responses:
        "201":
          description: Fan account created
        "409":
          description: Email already in use
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current staff user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user with memberships
  /auth/fan/me:
    get:
      tags: [Authentication]
      summary: Get current fan user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current fan user with fan profile

  # ── Retail / POS ──────────────────────────────────
...